fix: #1648 table tweaks buttons at the bottom
