+++ apps/trading/pages/portfolio/deposits-container.tsx
@@ -1,20 +1,15 @@
-import { AsyncRenderer, Button } from '@vegaprotocol/ui-toolkit';
-import { DepositsTable } from '@vegaprotocol/deposits';
+import { AsyncRenderer, Button, Dialog } from '@vegaprotocol/ui-toolkit';
+import { DepositContainer, DepositsTable } from '@vegaprotocol/deposits';
 import { useDeposits } from '@vegaprotocol/deposits';
 import { t } from '@vegaprotocol/react-helpers';
-import Link from 'next/link';
+import { useState } from 'react';
 
 export const DepositsContainer = () => {
   const { deposits, loading, error } = useDeposits();
+  const [depositDialog, setDepositDialog] = useState(false);
 
   return (
-    <div className="h-full grid grid-rows-[min-content_1fr]">
-      <header className="flex justify-between items-center p-4">
-        <h4 className="text-lg ">{t('Deposits')}</h4>
-        <Link href="/portfolio/deposit" passHref={true}>
-          <Button>Deposit</Button>
-        </Link>
-      </header>
+    <div className="h-full grid grid-rows-[1fr,min-content]">
       <div>
         <AsyncRenderer
           data={deposits}
@@ -25,6 +20,34 @@ export const DepositsContainer = () => {
           }}
         />
       </div>
+      <DepositDialog
+        depositDialog={depositDialog}
+        setDepositDialog={setDepositDialog}
+      />
+      <div className="w-full dark:bg-black bg-white absolute bottom-0 h-auto flex justify-end px-[11px] py-2">
+        <Button size="sm" onClick={() => setDepositDialog(true)}>
+          Deposit
+        </Button>
+      </div>
     </div>
+  );
+};
+
+export interface DepositDialogProps {
+  assetId?: string;
+  depositDialog: boolean;
+  setDepositDialog: (open: boolean) => void;
+}
+
+export const DepositDialog = ({
+  assetId,
+  depositDialog,
+  setDepositDialog,
+}: DepositDialogProps) => {
+  return (
+    <Dialog open={depositDialog} onChange={setDepositDialog}>
+      <h1 className="text-2xl mb-4">{t('Deposit')}</h1>
+      <DepositContainer assetId={assetId} />
+    </Dialog>
   );
 };
